requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
environs==11.0.0
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
idna==3.7
lxml==5.2.2
magic-filter==1.0.12
marshmallow==3.21.3
multidict==6.0.5
mutagen==1.47.0
packaging==24.1
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.3
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
yt-dlp==2024.7.9
